🐼 Complete expand category
